[MOV] moved test_compute_app_icon to test_website_event
`def test_compute_app_icon()` in `test_website.py`. added in forward port
https://github.com/odoo/odoo/pull/126643

it should be in `test_website_event.py`
